location + retour velo

+++ src/main/java/com/miage/vepick/repository/LocationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface LocationRepository extends CrudRepository<Location, Long> {
+public interface LocationRepository extends CrudRepository<Location, Integer> {
     
 }
